Refactored to address PR comments
